Zone and Room Protocol settings
firstly: pls check the details from the packdef.h file...

Secondly: i design the flow of going into the Five in Row Zone and
joining the empty play room.

Thirdly: the protocol about upgrading the room information, joining the
room, ready and starting Five in Row Game, and exiting the room.

Fourth: the first player join the room is the room owner and the other
is the participant.

Fifth: two line about exit. one is room owner exit, and the other is
room participant exit.

Come On !!!